Renamed EditQuote to EditQuotePage
