+++ app/containers/EditQuotePage/constants.js
