+++ app/containers/EditQuotePage/reducer.js
