+++ app/containers/EditQuotePage/tests/reducer.test.js
